requirements.txt
# 核心依赖
dash==2.14.1
plotly==5.17.0
networkx==3.2.1
pandas==2.1.4
numpy==1.24.3

# Web框架相关
flask==2.3.3
werkzeug==2.3.7

# 数据处理
json5==0.9.14
pyyaml==6.0.1

# 科学计算和可视化
scipy==1.11.4
matplotlib==3.8.2

# ⭐ 新增：NLP和文本处理
spacy>=3.4.0
nltk>=3.8
transformers>=4.20.0
jieba>=0.42.1
pkuseg>=0.0.25

# ⭐ 新增：机器学习
scikit-learn>=1.1.0
torch>=1.12.0
torchvision>=0.13.0
gensim>=4.2.0

# ⭐ 新增：文档处理
pdfplumber>=0.7.0
python-docx>=0.8.11
openpyxl>=3.0.10
pyPDF2>=2.10.0

# ⭐ 新增：邮件处理
email-validator>=1.3.0
imaplib2>=0.57

# ⭐ 新增：字符串匹配和相似度
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.20.0
fuzzy>=1.2.2

# ⭐ 新增：数据库和存储
pickle5>=0.0.12

# 开发和测试工具
pytest==7.4.3
pytest-cov==4.1.0
flake8==6.1.0
black==23.11.0

# 其他工具
click==8.1.7
colorama==0.4.6
tqdm>=4.64.0
requests>=2.28.0